requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
django-redis==5.4.0
flatbuffers==24.12.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
jsonschema==4.25.0
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.2
Pygments==2.18.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
pyzbar==0.1.9
requests==2.32.3
rich==13.9.4
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
django-ratelimit==4.1.0
django-redis==5.4.0
redis==5.0.1